refactor(#24): CarType 파일 SelectCarType으로 리네이밍 및 사용하지 않는 파일 삭제
- CarType -> SelectCarType 리네이밍: 함수명과 파일명을 일치시키기 위함
- 사용하지 않는 Select.tsx 파일 삭제
