--- client/src/pages/home/Select.tsx
@@ -1,5 +1,0 @@
-function Select() {
-
-}
-
-export default Select;
